+++ NEWS.md
@@ -1,6 +1,6 @@
 # teal.osprey 0.1.9
 
-### New feeatures
+### New features
 * Added new `teal` module `tm_g_heat_bygrade` for a heatmap by grade plot.
 * Added new `teal` module `tm_g_ae_sub` for an AE by subgroups plot.