+++ NEWS.md
@@ -1,3 +1,6 @@
+# teal.osprey 0.1.9.9000
+.
+
 # teal.osprey 0.1.9
 
 ### New features